+++ R/circa_single_mixed.R
@@ -12,6 +12,8 @@
 #' @param alpha_threshold The level of alpha for which the presence of rhythmicity is considered. Default is to \code{0.05}.
 #' @param nlme_control A list of control values for the estimation algorithm to replace the default values returned by the function nlme::nlmeControl. Defaults to an empty list.
 #' @param nlme_method A character string. If "REML" the model is fit by maximizing the restricted log-likelihood. If "ML" the log-likelihood is maximized. Defaults to "ML".
+#' @param weights An optional numeric vector of (fixed) weights internally passed to \code{nlme::nlme()} via \code{nlme::varPower()}.
+#' When present, the objective function is weighted least squares.
 #' @param suppress_all Logical. Set to \code{TRUE} to avoid seeing errors or messages during model fitting procedure. Default is \code{FALSE}. If \code{FALSE}, also runs \code{nlme()} with \code{verbose = TRUE}.
 #' @param timeout_n The upper limit for the model fitting attempts. Default is \code{10000}.
 #' @param return_figure Whether or not to return a ggplot graph of the rhythm and cosine model.
@@ -41,6 +43,14 @@
 #'   x = df, col_time = "time", col_outcome = "measure",
 #'   col_id = "id", randomeffects = c("k")
 #' )
+#'
+#' # with sample weights (arbitrary weights for demonstration)
+#' sw <- runif(n = nrow(df))
+#' out2 <- circa_single_mixed(
+#'   x = df, col_time = "time", col_outcome = "measure",
+#'   col_id = "id", randomeffects = c("k"), weights = sw
+#' )
+#'
 circa_single_mixed <- function(x,
                                col_time,
                                col_outcome,
@@ -50,6 +60,7 @@ circa_single_mixed <- function(x,
                                alpha_threshold = 0.05,
                                nlme_control = list(),
                                nlme_method = "ML",
+                               weights = NULL,
                                suppress_all = FALSE,
                                timeout_n = 10000,
                                return_figure = TRUE,
@@ -124,6 +135,13 @@ circa_single_mixed <- function(x,
     }
   }
 
+  if (!is.null(weights)) {
+    check_weights(x, weights)
+    x$weights <- weights
+  } else {
+    x$weights <- rep(1, nrow(x))
+  }
+
   success <- FALSE
   n <- 0
   form <- create_formula(main_params = controlVals$main_params, decay_params = controlVals$decay_params)$formula
@@ -141,7 +159,8 @@ circa_single_mixed <- function(x,
           start = unlist(start_list(outcome = x$measure, controlVals = controlVals)),
           control = nlme_control,
           method = nlme_method,
-          verbose = !suppress_all
+          verbose = !suppress_all,
+          weights = nlme::varPower(form = ~weights)
         )
       },
       silent = ifelse(suppress_all, TRUE, FALSE)